:recycle: rename package message to models
