+++ server_handler/internal/pubsub/pubsub_redis.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"log"
-	"server_handler/internal/message"
+	"server_handler/internal/models"
 
 	"github.com/redis/go-redis/v9"
 )
@@ -28,7 +28,7 @@ func (ps PubSubRedis) SendMessage(message, channel string) {
 
 // Listen pubsub for messages in format of the struct Message from internal/message
 // and send then to a channel from parameter
-func (ps PubSubRedis) ListenForMessages(channels []string, msgChannel chan message.Message) {
+func (ps PubSubRedis) ListenForMessages(channels []string, msgChannel chan models.Message) {
 	subscriber := ps.client.Subscribe(context.Background(), channels...)
 	_, err := subscriber.Receive(context.Background())
 	if err != nil {
@@ -39,7 +39,7 @@ func (ps PubSubRedis) ListenForMessages(channels []string, msgChannel chan messa
 
 	for msg := range ch {
 		log.Println("recebi aqui")
-		var m message.Message
+		var m models.Message
 
 		err := json.Unmarshal([]byte(msg.Payload), &m)
